Move server middleware configuration to related file
